update dynamics result service to use http client factory
